[task 0]: refactor original sbdd to support different drivers
For the purpose of this test task I've decided to go with a pseudo-driver model
that uses different sbdd implementations based on dymamic 'mode' variable. Basically
each mode corresponds to the task number, i.e. mode '0' creates vm disk and mode '1'
would work as a proxy over a real blkdevice.